+++ app/page.tsx
@@ -1,606 +1,75 @@
 "use client";
 
-import { useState, useEffect } from "react";
 import { Header } from "@/components/header";
 import { Sidebar } from "@/components/sidebar";
 import { ChatCanvas } from "@/components/chat-canvas";
 import { SessionModal } from "@/components/session-modal";
-import { ResumeUpload } from "@/components/resume-upload";
-import { InterviewCustomization } from "@/components/interview-customization";
-import type { Character, InterviewType, Message, SessionData, ResumeData, InterviewCustomization as CustomizationType } from "@/types";
-import { createEmptyResumeData, validateResumeData } from "@/types";
+import { WelcomeScreen } from "@/components/WelcomeScreen";
 import { Loader2 } from "lucide-react";
-import type { FaceMetrics } from "@/lib/hooks/useFaceDetection";
-
-const characters: Character[] = [
-  {
-    id: "jane",
-    name: "Jane Doe",
-    role: "Tech Lead",
-    company: "TechVision Corp",
-    years: 10,
-    focus: "frontend architecture and scalable web applications",
-    avatar: "/placeholder.svg?height=48&width=48",
-    avatarAnimated: "/avatars/female_laptop.jpg",
-    description:
-      "Experienced frontend architect with 8+ years at top tech companies",
-  },
-  {
-    id: "mike",
-    name: "Mike Chen",
-    role: "HR Manager",
-    company: "InnovateTech Solutions",
-    years: 8,
-    focus:
-      "building high-performing teams in AI-driven applications and product development",
-    avatar: "/placeholder.svg?height=48&width=48",
-    avatarAnimated: "/avatars/male_talking.jpg",
-    description: "People-focused leader specializing in behavioral interviews",
-  },
-  {
-    id: "sarah",
-    name: "Sarah Wilson",
-    role: "Product Manager",
-    company: "NextGen Products",
-    years: 6,
-    focus: "strategic product launches and cross-functional team leadership",
-    avatar: "/placeholder.svg?height=48&width=48",
-    avatarAnimated: "/avatars/female_mic.jpg",
-    description: "Strategic thinker with expertise in case study interviews",
-  },
-];
-
-const interviewTypes: InterviewType[] = [
-  {
-    id: "technical",
-    name: "Technical",
-    icon: "💻",
-    description: "Coding challenges and system design",
-  },
-  {
-    id: "behavioral",
-    name: "Behavioral",
-    icon: "🤝",
-    description: "Situational and experience-based questions",
-  },
-  {
-    id: "case-study",
-    name: "Case Study",
-    icon: "📊",
-    description: "Problem-solving and analytical thinking",
-  },
-];
+import { characters } from "@/data/characters";
+import { interviewTypes } from "@/data/interviewTypes";
+import { useInterviewLogic } from "@/hooks/useInterviewLogic";
+import { useSidebarState } from "@/hooks/useSidebarState";
 
 export default function MockInterviewAI() {
-  const [selectedCharacter, setSelectedCharacter] = useState<Character | null>(
-    null
-  );
-  const [selectedType, setSelectedType] = useState<InterviewType | null>(null);
-  const [messages, setMessages] = useState<Message[]>([]);
-  const [isRecording, setIsRecording] = useState(false);
-  const [sessionComplete, setSessionComplete] = useState(false);
-  const [sessionData, setSessionData] = useState<SessionData | null>(null);
-  const [currentQuestion, setCurrentQuestion] = useState(1);
-  const [isDarkMode, setIsDarkMode] = useState(false);
-  const [chatHistory, setChatHistory] = useState<Message[]>([]);
-  const [analysisLoading, setAnalysisLoading] = useState(false);
-  const [interviewStarted, setInterviewStarted] = useState(false);
-  const [selectedLanguage, setSelectedLanguage] = useState("en");
-  const [faceMetricsHistory, setFaceMetricsHistory] = useState<FaceMetrics[]>([]);
-  const [resumeData, setResumeData] = useState<ResumeData | null>(null);
-  const [isResumeUploaded, setIsResumeUploaded] = useState(false);
-  const [interviewCustomization, setInterviewCustomization] = useState<CustomizationType>({
-    difficulty: 'moderate',
-    topicFocus: 'mixed',
-    purpose: 'general'
-  });
-  const [sidebarOpen, setSidebarOpen] = useState(false);
-
-  // Simple translation object (you can expand this)
-  const translations = {
-    en: {
-      welcomeTitle: "Welcome to MockInterviewAI",
-      welcomeSubtitle:
-        "Select your interviewer and interview type from the sidebar, then click Start Interview to begin your mock session.",
-      startButton: "Start Interview",
-    },
-    es: {
-      welcomeTitle: "Bienvenido a MockInterviewAI",
-      welcomeSubtitle:
-        "Selecciona tu entrevistador y tipo de entrevista en la barra lateral, luego haz clic en Iniciar entrevista para comenzar tu sesión de simulación.",
-      startButton: "Iniciar Entrevista",
-    },
-    fr: {
-      welcomeTitle: "Bienvenue sur MockInterviewAI",
-      welcomeSubtitle:
-        "Sélectionnez votre intervieweur et type d'entretien dans la barre latérale, puis cliquez sur Démarrer l'entretien pour commencer votre session simulée.",
-      startButton: "Démarrer l'Entretien",
-    },
-    hi: {
-      welcomeTitle: "Welcome to MockInterviewAI (Hindi)",
-      welcomeSubtitle:
-        "Select your interviewer and interview type from the sidebar, then click Start Interview to begin your mock session. (Hindi)",
-      startButton: "Start Interview (Hindi)",
-    },
-  };
-
-  // Function to get translated text
-  const t = (key: keyof typeof translations.en) => {
-    return (
-      translations[selectedLanguage as keyof typeof translations][key] ||
-      translations.en[key]
-    );
-  };
-
-  // Store chat history in localStorage
-  useEffect(() => {
-    if (messages.length > 0) {
-      localStorage.setItem("chatHistory", JSON.stringify(messages));
-      setChatHistory(messages);
-    }
-  }, [messages]);
-
-  // Load chat history on component mount
-  useEffect(() => {
-    const savedHistory = localStorage.getItem("chatHistory");
-    if (savedHistory) {
-      const parsedHistory = JSON.parse(savedHistory);
-      setChatHistory(parsedHistory);
-    }
-  }, []);
-
-  const getInterviewerIntroduction = (
-    character: Character,
-    type: InterviewType
-  ) => {
-    const difficultyText = interviewCustomization.difficulty === 'beginner' ? 'beginner-friendly' : 
-                          interviewCustomization.difficulty === 'advanced' ? 'advanced' : 'moderate';
-    const topicText = interviewCustomization.topicFocus === 'mixed' ? 'various topics' : 
-                     `${interviewCustomization.topicFocus} topics`;
-    const purposeText = interviewCustomization.purpose === 'intern' ? 'internship' : 
-                       interviewCustomization.purpose === 'placement' ? 'placement' : 'general';
-
-    if (resumeData && resumeData.name) {
-      return `Hi ${resumeData.name}, I'm ${character.name}, ${
-        character.role
-      } for the AI team. I've reviewed your resume and I'll be conducting a ${difficultyText} ${type.name.toLowerCase()} interview focused on ${topicText} for ${purposeText} preparation. Let's start with a brief introduction - please tell me about yourself and your background.`;
-    }
-    return `Hi, I'm ${character.name}, ${
-      character.role
-    } for the AI team. I'll be conducting a ${difficultyText} ${type.name.toLowerCase()} interview focused on ${topicText} for ${purposeText} preparation. Let's start with a brief introduction - please tell me about yourself and your background.`;
-  };
-
-  const initializeSession = async (
-    character: Character,
-    type: InterviewType
-  ) => {
-    const welcomeMessage = getInterviewerIntroduction(character, type);
-    setMessages([
-      {
-        id: "1",
-        type: "interviewer",
-        content: welcomeMessage,
-        timestamp: new Date(),
-        character: character,
-        hasAudio: false,
-      },
-    ]);
-    setCurrentQuestion(1);
-
-    // try {
-    //   // Include language in API call
-    //   const response = await fetch("/api/process-message", {
-    //     method: "POST",
-    //     headers: {
-    //       "Content-Type": "application/json",
-    //     },
-    //     body: JSON.stringify({
-    //       message: welcomeMessage,
-    //       character: character,
-    //       interviewType: type,
-    //       currentQuestion: 1,
-    //       language: selectedLanguage,
-    //     }),
-    //   });
-
-    //   const data = await response.json();
-
-    //   if (data.success) {
-    //     const aiMessage: Message = {
-    //       id: (Date.now() + 1).toString(),
-    //       type: "interviewer",
-    //       content: data.processedMessage,
-    //       timestamp: new Date(),
-    //       character: character,
-    //       hasAudio: true,
-    //       audioFile: data.audioFile,
-    //     };
-
-    //     setMessages((prev) => [...prev, aiMessage]);
-    //     setCurrentQuestion((prev) => prev + 1);
-
-    //     if (currentQuestion >= 5) {
-    //       // Generate final feedback using Gemini
-    //       const feedbackPrompt = `As ${character.name}, a ${character.role}, provide a comprehensive interview feedback based on the candidate's responses.
-    //       Include:
-    //       1. Overall score (out of 10)
-    //       2. Key strengths (3 points)
-    //       3. Areas for improvement (3 points)
-    //       4. Interview duration
-    //       5. Number of questions answered
-          
-    //       Format the response as a JSON object with these fields:
-    //       {
-    //         "score": number,
-    //         "strengths": string[],
-    //         "improvements": string[],
-    //         "duration": string,
-    //         "questionsAnswered": number
-    //       }`;
-
-    //       const feedbackResponse = await fetch("/api/process-message", {
-    //         method: "POST",
-    //         headers: {
-    //           "Content-Type": "application/json",
-    //         },
-    //         body: JSON.stringify({
-    //           message: feedbackPrompt,
-    //           character: character,
-    //           interviewType: type,
-    //           isFeedback: true,
-    //           language: selectedLanguage,
-    //         }),
-    //       });
-
-    //       const feedbackData = await feedbackResponse.json();
-
-    //       if (feedbackData.success) {
-    //         try {
-    //           const cleaned = feedbackData.processedMessage
-    //             .replace(/```json|```/g, "")
-    //             .trim();
-    //           const feedback = JSON.parse(cleaned);
-    //           setSessionData(feedback);
-    //           setSessionComplete(true);
-    //         } catch (error) {
-    //           console.error("Error parsing feedback:", error);
-    //         }
-    //       }
-    //     }
-    //   }
-    // } catch (error) {
-    //   console.error("Error processing message:", error);
-    // }
-  };
-
-  const handleCharacterSelect = async (character: Character) => {
-    setSelectedCharacter(character);
-  };
-
-  const handleTypeSelect = async (type: InterviewType) => {
-    setSelectedType(type);
-  };
-
-  const handleVoiceInput = async (transcript: string) => {
-    if (!selectedCharacter || !selectedType) return;
-
-    const userMessage: Message = {
-      id: Date.now().toString(),
-      type: "user",
-      content: transcript,
-      timestamp: new Date(),
-    };
-
-    setMessages((prev) => [...prev, userMessage]);
-
-    try {
-      // Create a prompt for Gemini based on the character, interview type, and resume data
-      let prompt = `You are ${selectedCharacter.name}, a ${
-        selectedCharacter.role
-      } conducting a ${selectedType.name.toLowerCase()} interview. 
-      The candidate just said: "${transcript}"
-      
-      Based on this response and the interview context:
-      - Character: ${selectedCharacter.name} (${selectedCharacter.role})
-      - Interview Type: ${selectedType.name}
-      - Current Question Number: ${currentQuestion} of 5
-      - Target Language: ${selectedLanguage}
-      - Difficulty Level: ${interviewCustomization.difficulty}
-      - Topic Focus: ${interviewCustomization.topicFocus}
-      - Interview Purpose: ${interviewCustomization.purpose}`;
-
-      // Add resume context if available
-      if (resumeData && isResumeUploaded) {
-        prompt += `
-      
-      Candidate's Resume Information:
-      - Name: ${resumeData.name}
-      - Summary: ${resumeData.summary}
-      - Experience: ${resumeData.experience.map(exp => `${exp.title} at ${exp.company} (${exp.duration})`).join(', ')}
-      - Skills: ${resumeData.skills.join(', ')}
-      - Projects: ${resumeData.projects.map(proj => proj.name).join(', ')}
-      - Achievements: ${resumeData.achievements.join(', ')}`;
-      }
-
-      prompt += `
-      
-      Generate a relevant follow-up question that:
-      1. Is specific to the ${selectedType.name.toLowerCase()} interview type
-      2. Shows expertise in ${selectedCharacter.role} role
-      3. Builds upon the candidate's previous response
-      4. Helps assess their skills and experience
-      5. Is in ${selectedLanguage}
-      6. Matches the ${interviewCustomization.difficulty} difficulty level
-      7. Focuses on ${interviewCustomization.topicFocus} topics
-      8. Is appropriate for ${interviewCustomization.purpose} interviews`;
-
-      // Add resume-specific instructions
-      if (resumeData && isResumeUploaded) {
-        prompt += `
-      9. References specific details from their resume (experience, projects, skills, or achievements)
-      10. Asks about their actual work experience, projects, or technical skills mentioned in their resume`;
-      }
-
-      prompt += `
-      
-      Also, analyze your mood as the interviewer based on the candidate's response. Consider:
-      - The quality and relevance of their answer
-      - Their communication skills
-      - Their level of engagement
-      - Their technical/professional knowledge shown
-      
-      Format your response as a JSON object with these fields:
-      {
-        "question": "your follow-up question",
-        "mood": {
-          "value": number (0-100, where 0 is very negative and 100 is very positive),
-          "label": "one word mood label (e.g. Positive, Neutral, Concerned)",
-          "emoji": "appropriate emoji for the mood"
-        }
-      }
-      
-      Keep the question concise and professional.`;
-
-      const response = await fetch("/api/process-message", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          message: prompt,
-          character: selectedCharacter,
-          interviewType: selectedType,
-          currentQuestion,
-          language: selectedLanguage,
-        }),
-      });
-
-      const data = await response.json();
-
-      if (data.success) {
-        try {
-          const cleaned = data.processedMessage.replace(/```json|```/g, "").trim();
-          const response = JSON.parse(cleaned);
-          
-          const aiMessage: Message = {
-            id: (Date.now() + 1).toString(),
-            type: "interviewer",
-            content: response.question,
-            timestamp: new Date(),
-            character: selectedCharacter,
-            hasAudio: true,
-            audioFile: data.audioFile,
-            mood: response.mood
-          };
-
-          setMessages((prev) => [...prev, aiMessage]);
-          setCurrentQuestion((prev) => prev + 1);
-
-          // if (currentQuestion >= 5) {
-          //   // Generate final feedback using Gemini
-          //   const feedbackPrompt = `As ${selectedCharacter.name}, a ${selectedCharacter.role}, provide a comprehensive interview feedback based on the candidate's responses.
-          //   Include:
-          //   1. Overall score (out of 10)
-          //   2. Key strengths (3 points)
-          //   3. Areas for improvement (3 points)
-          //   4. Interview duration
-          //   5. Number of questions answered
-            
-          //   Format the response as a JSON object with these fields:
-          //   {
-          //     "score": number,
-          //     "strengths": string[],
-          //     "improvements": string[],
-          //     "duration": string,
-          //     "questionsAnswered": number
-          //   }`;
-
-          //   const feedbackResponse = await fetch("/api/process-message", {
-          //     method: "POST",
-          //     headers: {
-          //       "Content-Type": "application/json",
-          //     },
-          //     body: JSON.stringify({
-          //       message: feedbackPrompt,
-          //       character: selectedCharacter,
-          //       interviewType: selectedType,
-          //       isFeedback: true,
-          //       language: selectedLanguage,
-          //     }),
-          //   });
-
-          //   const feedbackData = await feedbackResponse.json();
-
-          //   if (feedbackData.success) {
-          //     try {
-          //       const cleaned = feedbackData.processedMessage
-          //         .replace(/```json|```/g, "")
-          //         .trim();
-          //       const feedback = JSON.parse(cleaned);
-          //       setSessionData(feedback);
-          //       setSessionComplete(true);
-          //     } catch (error) {
-          //       console.error("Error parsing feedback:", error);
-          //     }
-          //   }
-          // }
-        } catch (error) {
-          console.error("Error parsing response:", error);
-        }
-      }
-    } catch (error) {
-      console.error("Error processing message:", error);
-    }
-  };
-
-  const handleFaceMetricsUpdate = (metrics: FaceMetrics) => {
-    setFaceMetricsHistory(prev => [...prev, metrics]);
-  };
-
-  const calculateFaceMetricsSummary = (metrics: FaceMetrics[]): SessionData['faceMetrics'] => {
-    if (metrics.length === 0) return undefined;
-
-    const averageConfidence = metrics.reduce((sum, m) => sum + m.confidence, 0) / metrics.length;
+  const {
+    // State
+    selectedCharacter,
+    selectedType,
+    messages,
+    isRecording,
+    sessionComplete,
+    sessionData,
+    currentQuestion,
+    isDarkMode,
+    chatHistory,
+    analysisLoading,
+    interviewStarted,
+    selectedLanguage,
+    faceMetricsHistory,
+    resumeData,
+    isResumeUploaded,
+    interviewCustomization,
     
-    const expressionCounts: Record<string, number> = {};
-    metrics.forEach(m => {
-      const dominantExpression = Object.entries(m.expressions).reduce(
-        (a, b) => (a[1] > b[1] ? a : b)
-      )[0];
-      expressionCounts[dominantExpression] = (expressionCounts[dominantExpression] || 0) + 1;
-    });
-
-    const eyeContactPercentage = metrics.filter(m => m.eyeContact).length / metrics.length;
-
-    return {
-      averageConfidence,
-      dominantExpressions: expressionCounts,
-      eyeContactPercentage,
-    };
-  };
-
-  const handleAnalysis = async () => {
-    if (!selectedCharacter || !selectedType || messages.length === 0) return;
-    setAnalysisLoading(true);
-    try {
-      let analysisPrompt = `As ${selectedCharacter.name}, a ${
-        selectedCharacter.role
-      }, analyze this interview session and provide detailed feedback in ${selectedLanguage}.\n\nInterview Type: ${
-        selectedType.name
-      }\nInterview Customization:
-- Difficulty Level: ${interviewCustomization.difficulty}
-- Topic Focus: ${interviewCustomization.topicFocus}
-- Interview Purpose: ${interviewCustomization.purpose}
-\nChat History:\n${messages
-        .map(
-          (msg) =>
-            `${msg.type === "user" ? "Candidate" : "Interviewer"}: ${
-              msg.content
-            }`
-        )
-        .join(
-          "\n"
-        )}`;
-
-      // Add resume context if available
-      if (resumeData && isResumeUploaded) {
-        analysisPrompt += `\n\nCandidate's Resume Information:
-- Name: ${resumeData.name}
-- Summary: ${resumeData.summary}
-- Experience: ${resumeData.experience.map(exp => `${exp.title} at ${exp.company} (${exp.duration})`).join(', ')}
-- Skills: ${resumeData.skills.join(', ')}
-- Projects: ${resumeData.projects.map(proj => proj.name).join(', ')}
-- Achievements: ${resumeData.achievements.join(', ')}`;
-      }
-
-      analysisPrompt += `\n\nPlease provide:\n1. Overall rating (out of 5 stars)\n2. Key strengths (3 points)\n3. Areas for improvement (3 points)\n4. Communication skills assessment\n5. Technical/Professional knowledge evaluation
-6. Specific recommendations for improvement`;
-
-      // Add resume-specific analysis instructions
-      if (resumeData && isResumeUploaded) {
-        analysisPrompt += `\n7. How well their responses aligned with their resume experience
-8. Whether they effectively communicated their actual skills and achievements
-9. Suggestions for better connecting their responses to their documented experience`;
-      }
-      
-      analysisPrompt += `\n\nFormat the response as a JSON object with these fields, ensuring all text is in ${selectedLanguage}:\n{\n  "rating": number,\n  "strengths": string[],\n  "improvements": string[],\n  "communication": string,\n  "knowledge": string,\n  "recommendations": string[]\n}`;
-
-      const response = await fetch("/api/process-message", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          message: analysisPrompt,
-          character: selectedCharacter,
-          interviewType: selectedType,
-          isAnalysis: true,
-          language: selectedLanguage,
-        }),
-      });
-      const data = await response.json();
-      if (data.success) {
-        try {
-          const cleaned = data.processedMessage
-            .replace(/```json|```/g, "")
-            .trim();
-          const analysis = JSON.parse(cleaned);
-          
-          // Add face metrics to the analysis
-          analysis.faceMetrics = calculateFaceMetricsSummary(faceMetricsHistory);
-          
-          setSessionData(analysis);
-          setSessionComplete(true);
-        } catch (error) {
-          console.error("Error parsing analysis:", error);
-        }
-      }
-    } catch (error) {
-      console.error("Error generating analysis:", error);
-    } finally {
-      setAnalysisLoading(false);
-    }
-  };
-
-  const handleResumeExtracted = (data: ResumeData) => {
-    setResumeData(data);
-    const validation = validateResumeData(data);
-    setIsResumeUploaded(validation.isValid);
+    // Actions
+    setIsDarkMode,
+    setSelectedLanguage,
+    setInterviewCustomization,
+    setAnalysisLoading,
+    setInterviewStarted,
+    setIsRecording,
     
-    if (!validation.isValid) {
-      console.warn('Resume validation errors:', validation.errors);
-    }
-    if (validation.warnings.length > 0) {
-      console.warn('Resume validation warnings:', validation.warnings);
+    // Functions
+    handleCharacterSelect,
+    handleTypeSelect,
+    handleVoiceInput,
+    handleFaceMetricsUpdate,
+    handleAnalysis,
+    handleResumeExtracted,
+    initializeSession,
+    resetInterviewState,
+    handleSessionModalClose,
+    handleSessionModalRetry,
+  } = useInterviewLogic();
+
+  const { isSidebarOpen, toggleSidebar } = useSidebarState();
+
+  const handleStartInterview = async () => {
+    setInterviewStarted(true);
+    if (selectedCharacter && selectedType) {
+      await initializeSession(selectedCharacter, selectedType);
     }
   };
 
-  // Helper function to clear resume state
-  const clearResumeState = () => {
-    setResumeData(null);
-    setIsResumeUploaded(false);
-  };
-
-  // Helper function to reset interview state
-  const resetInterviewState = () => {
-    setMessages([]);
-    setCurrentQuestion(1);
-    setInterviewStarted(false);
-    setFaceMetricsHistory([]);
-    clearResumeState();
-    setSidebarOpen(false);
-    setInterviewCustomization({
-      difficulty: 'moderate',
-      topicFocus: 'mixed',
-      purpose: 'general'
-    });
+  const handleViewHistory = () => {
+    console.log(chatHistory);
   };
 
   // Show Start Interview button if not started
   if (!interviewStarted) {
     return (
       <div
-        className={`min-h-screen pt-16 transition-colors duration-300 ${  // Added pt-16 here
+        className={`min-h-screen pt-16 transition-colors duration-300 ${
           isDarkMode ? "dark bg-gray-900" : "bg-[#F7F7F9]"
         }`}
       >
@@ -611,7 +80,6 @@ export default function MockInterviewAI() {
           onToggleDarkMode={() => setIsDarkMode(!isDarkMode)}
           selectedLanguage={selectedLanguage}
           onLanguageSelect={setSelectedLanguage}
-          onToggleSidebar={() => setSidebarOpen(!sidebarOpen)}
         />
         <div className="flex h-[calc(100vh-64px)]">
           <Sidebar
@@ -621,98 +89,24 @@ export default function MockInterviewAI() {
             selectedType={selectedType}
             onCharacterSelect={handleCharacterSelect}
             onTypeSelect={handleTypeSelect}
-            onResetInterview={() => {
-              resetInterviewState();
-            }}
-            onNewInterview={() => {
-              resetInterviewState();
-            }}
-            onViewHistory={() => {
-              // You can implement a modal or navigation to view chatHistory if needed
-              // For now, just log it
-              console.log(chatHistory);
-            }}
+            onResetInterview={resetInterviewState}
+            onNewInterview={resetInterviewState}
+            onViewHistory={handleViewHistory}
             messages={messages}
-            isOpen={sidebarOpen}
-            onToggle={() => setSidebarOpen(!sidebarOpen)}
+            isOpen={isSidebarOpen}
+            onToggle={toggleSidebar}
           />
-          <div className="flex-1 flex flex-col items-center justify-center p-4 lg:p-8 lg:ml-80">
-            <div className="max-w-6xl w-full space-y-6 lg:space-y-8">
-              <div className="text-center">
-                <h1 className="text-2xl lg:text-3xl font-bold text-gray-900 dark:text-white mb-4 lg:mb-6 mt-8 lg:mt-12">
-                  {t("welcomeTitle")}
-                </h1>
-                <p className="text-base lg:text-lg text-gray-600 dark:text-gray-300 max-w-2xl mx-auto text-center mb-6 lg:mb-8">
-                  {t("welcomeSubtitle")}
-                </p>
-              </div>
-
-              <div className="grid grid-cols-1 lg:grid-cols-2 xl:grid-cols-3 gap-6 lg:gap-8 items-start">
-                {/* Resume Upload Section */}
-                <div className="flex flex-col items-center">
-                  <ResumeUpload
-                    onResumeExtracted={handleResumeExtracted}
-                    isUploaded={isResumeUploaded}
-                    resumeData={resumeData}
-                  />
-                </div>
-
-                {/* Interview Customization Section */}
-                <div className="flex flex-col items-center lg:col-span-1 xl:col-span-1">
-                  <InterviewCustomization
-                    customization={interviewCustomization}
-                    onCustomizationChange={setInterviewCustomization}
-                  />
-                </div>
-
-                {/* Interview Setup Section */}
-                <div className="flex flex-col justify-center space-y-4 lg:space-y-6 lg:col-span-1 xl:col-span-1">
-                  <div className="text-center">
-                    <h2 className="text-xl lg:text-2xl font-semibold text-gray-900 dark:text-white mb-3 lg:mb-4">
-                      Interview Setup
-                    </h2>
-                    <p className="text-sm lg:text-base text-gray-600 dark:text-gray-300 mb-6 lg:mb-8 max-w-md mx-auto">
-                      {isResumeUploaded 
-                        ? "Great! Your resume has been uploaded. The interviewer will ask personalized questions based on your experience."
-                        : "Upload your resume for personalized questions, or proceed without it for general questions."
-                      }
-                    </p>
-                  </div>
-
-                  <div className="space-y-3 lg:space-y-4">
-                    {!selectedCharacter && (
-                      <div className="p-3 lg:p-4 bg-yellow-50 dark:bg-yellow-900/20 border border-yellow-200 dark:border-yellow-800 rounded-lg">
-                        <p className="text-xs lg:text-sm text-yellow-800 dark:text-yellow-200">
-                          Please select an interviewer from the sidebar
-                        </p>
-                      </div>
-                    )}
-                    
-                    {!selectedType && (
-                      <div className="p-3 lg:p-4 bg-yellow-50 dark:bg-yellow-900/20 border border-yellow-200 dark:border-yellow-800 rounded-lg">
-                        <p className="text-xs lg:text-sm text-yellow-800 dark:text-yellow-200">
-                          Please select an interview type from the sidebar
-                        </p>
-                      </div>
-                    )}
-                  </div>
-
-                  <button
-                    className="px-6 lg:px-8 py-3 lg:py-4 bg-[#E07A5F] text-white text-base lg:text-lg font-semibold rounded-xl shadow-lg hover:bg-[#E07A5F]/90 transition disabled:opacity-50 disabled:cursor-not-allowed"
-                    disabled={!selectedCharacter || !selectedType}
-                    onClick={async () => {
-                      setInterviewStarted(true);
-                      if (selectedCharacter && selectedType) {
-                        await initializeSession(selectedCharacter, selectedType);
-                      }
-                    }}
-                  >
-                    {t("startButton")}
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
+          <WelcomeScreen
+            selectedCharacter={selectedCharacter}
+            selectedType={selectedType}
+            selectedLanguage={selectedLanguage}
+            isResumeUploaded={isResumeUploaded}
+            resumeData={resumeData}
+            interviewCustomization={interviewCustomization}
+            onResumeExtracted={handleResumeExtracted}
+            onCustomizationChange={setInterviewCustomization}
+            onStartInterview={handleStartInterview}
+          />
         </div>
       </div>
     );
@@ -720,7 +114,7 @@ export default function MockInterviewAI() {
 
   return (
     <div
-      className={`min-h-screen pt-16 transition-colors duration-300 ${  // Added pt-16 here
+      className={`min-h-screen pt-16 transition-colors duration-300 ${
         isDarkMode ? "dark bg-gray-900" : "bg-[#F7F7F9]"
       }`}
     >
@@ -731,7 +125,6 @@ export default function MockInterviewAI() {
         onToggleDarkMode={() => setIsDarkMode(!isDarkMode)}
         selectedLanguage={selectedLanguage}
         onLanguageSelect={setSelectedLanguage}
-        onToggleSidebar={() => setSidebarOpen(!sidebarOpen)}
       />
 
       <div className="flex h-[calc(100vh-64px)]">
@@ -742,20 +135,12 @@ export default function MockInterviewAI() {
           selectedType={selectedType}
           onCharacterSelect={handleCharacterSelect}
           onTypeSelect={handleTypeSelect}
-          onResetInterview={() => {
-            resetInterviewState();
-          }}
-          onNewInterview={() => {
-            resetInterviewState();
-          }}
-          onViewHistory={() => {
-            // You can implement a modal or navigation to view chatHistory if needed
-            // For now, just log it
-            console.log(chatHistory);
-          }}
+          onResetInterview={resetInterviewState}
+          onNewInterview={resetInterviewState}
+          onViewHistory={handleViewHistory}
           messages={messages}
-          isOpen={sidebarOpen}
-          onToggle={() => setSidebarOpen(!sidebarOpen)}
+          isOpen={isSidebarOpen}
+          onToggle={toggleSidebar}
         />
 
         <ChatCanvas
@@ -785,17 +170,12 @@ export default function MockInterviewAI() {
           </div>
         </div>
       )}
+      
       {sessionComplete && sessionData && !analysisLoading && (
         <SessionModal
           sessionData={sessionData}
-          onClose={() => setSessionComplete(false)}
-          onRetry={() => {
-            setSessionComplete(false);
-            resetInterviewState();
-            if (selectedCharacter && selectedType) {
-              initializeSession(selectedCharacter, selectedType);
-            }
-          }}
+          onClose={handleSessionModalClose}
+          onRetry={handleSessionModalRetry}
         />
       )}
     </div>
